Able access the Firebase Data base- Take A break
